backend/app/requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain-google-genai
langchain-community
pytube
markdown
python-dotenv
